Moved class into Scala Package
